refactor: 문자열 분해로직 클래스로 분리
